MAINT: Rename folders
