+++ src/fmu/dataio/external_interfaces/schema_validation_interface.py
